Added `migration` export for deprecated design tokens
