+++ packages/adaptive-ui-explorer/src/components/swatch.ts
@@ -1,5 +1,6 @@
 import { Swatch } from "@adaptive-web/adaptive-ui";
-import { fillColor, neutralForegroundHint } from "@adaptive-web/adaptive-ui/reference";
+import { neutralForegroundHint } from "@adaptive-web/adaptive-ui/migration";
+import { fillColor } from "@adaptive-web/adaptive-ui/reference";
 import { contrastRatio, parseColor } from "@microsoft/fast-colors";
 import {
     attr,
