+++ packages/adaptive-ui/src/migration/index.ts
@@ -1,0 +1,4 @@
+export * from "./appearance.js";
+export * from "./color.js";
+export * from "./type-ramp.js";
+export * from "./type.js";
