+++ packages/adaptive-ui/src/migration/type.ts
@@ -1,0 +1,4 @@
+import { bodyFontFamily } from "../design-tokens/type.js";
+
+/** @public @deprecated Renamed to `bodyFontFamily` */
+export const bodyFont = bodyFontFamily;
