+++ packages/adaptive-ui/src/reference.ts
@@ -1,2 +1,1 @@
 export * from "./design-tokens/index.js";
-export * from "./type/index.js";
