--- packages/adaptive-ui/src/type/index.ts
@@ -1,1 +1,0 @@
-export * from "./type-ramp.js";
