+++ packages/adaptive-web-components/src/components/button/button.styles.ts
@@ -1,7 +1,7 @@
 import { css, ElementStyles } from "@microsoft/fast-element";
 import {
     focusStrokeOuter,
-    focusStrokeWidth,
+    focusStrokeThickness,
 } from "@adaptive-web/adaptive-ui/reference";
 
 /**
@@ -57,6 +57,6 @@ export const aestheticStyles: ElementStyles = css`
     }
 
     .control:focus-visible {
-        outline: calc(${focusStrokeWidth} * 1px) solid ${focusStrokeOuter};
+        outline: ${focusStrokeThickness} solid ${focusStrokeOuter};
     }
 `;
