+++ packages/adaptive-web-components/src/components/tabs/tabs.styles.ts
@@ -1,7 +1,7 @@
 import {
     accentStrokeReadableRest,
-    controlCornerRadius,
-    focusStrokeWidth,
+    cornerRadiusControl,
+    focusStrokeThickness,
 } from "@adaptive-web/adaptive-ui/reference";
 import { css, ElementStyles } from "@microsoft/fast-element";
 import { heightNumber } from "../../styles/index.js";
@@ -75,18 +75,18 @@ export const aestheticStyles: ElementStyles = css`
         content: "";
         display: block;
         position: absolute;
-        border-radius: calc(${controlCornerRadius} * 1px);
+        border-radius: ${cornerRadiusControl};
         background: ${accentStrokeReadableRest};
     }
 
     :host([orientation="horizontal"]) ::slotted([role="tab"][aria-selected="true"])::after {
-        bottom: calc(${focusStrokeWidth} * 1px);
+        bottom: ${focusStrokeThickness};
         width: 20px;
         height: 3px;
     }
 
     :host([orientation="vertical"]) ::slotted([role="tab"][aria-selected="true"])::after {
-        left: calc(${focusStrokeWidth} * 1px);
+        left: ${focusStrokeThickness};
         width: 3px;
         height: calc((${heightNumber} / 2) * 1px);
     }
