+++ packages/adaptive-web-components/src/components/toolbar/toolbar.styles.ts
@@ -1,4 +1,5 @@
-import { focusStrokeOuter, focusStrokeWidth, neutralForegroundRest } from "@adaptive-web/adaptive-ui/reference";
+import { neutralForegroundRest } from "@adaptive-web/adaptive-ui/migration";
+import { focusStrokeOuter, focusStrokeThickness } from "@adaptive-web/adaptive-ui/reference";
 import { css, ElementStyles } from "@microsoft/fast-element";
 
 /**
@@ -46,7 +47,7 @@ export const aestheticStyles: ElementStyles = css`
     }
 
     :host(:focus-visible) {
-        outline: calc(${focusStrokeWidth} * 1px) solid ${focusStrokeOuter};
+        outline: ${focusStrokeThickness} solid ${focusStrokeOuter};
     }
 
     .positioning-region {
